--- automation/src/test/scala/org/broadinstitute/dsde/test/api/orch/AsyncTSVImportSpec
@@ -1,76 +1,0 @@
-package org.broadinstitute.dsde.test.api.orch
-
-
-import java.util.UUID
-
-import akka.http.scaladsl.model.{HttpResponse, StatusCodes}
-import akka.http.scaladsl.unmarshalling.Unmarshal
-import org.broadinstitute.dsde.rawls.model.EntityTypeMetadata
-import org.broadinstitute.dsde.rawls.model.WorkspaceJsonSupport.EntityTypeMetadataFormat
-import org.broadinstitute.dsde.workbench.auth.AuthToken
-import org.broadinstitute.dsde.workbench.config.{Credentials, ServiceTestConfig, UserPool}
-import org.broadinstitute.dsde.workbench.fixture.{BillingFixtures, WorkspaceFixtures}
-import org.broadinstitute.dsde.workbench.model.ErrorReport
-import org.broadinstitute.dsde.workbench.model.ErrorReportJsonSupport.ErrorReportFormat
-import org.broadinstitute.dsde.workbench.service.{AclEntry, Orchestration, RestException, WorkspaceAccessLevel}
-import org.parboiled.common.FileUtils
-import org.scalatest.OptionValues._
-import org.scalatest.concurrent.{Eventually, ScalaFutures}
-import org.scalatest.time.{Seconds, Span}
-import org.scalatest.{FreeSpec, Matchers}
-import spray.json._
-
-
-import scala.io.Source
-
-class AsyncTSVImportSpec extends FreeSpec with Matchers with Eventually with ScalaFutures
-  with BillingFixtures with WorkspaceFixtures with Orchestration {
-
-  val owner: Credentials = UserPool.chooseProjectOwner
-  val ownerAuthToken: AuthToken = owner.makeAuthToken()
-
-  final implicit override val patienceConfig: PatienceConfig = PatienceConfig(timeout = scaled(Span(300, Seconds)), interval = scaled(Span(2, Seconds)))
-
-  private val testPayload = FileUtils.readAllTextFromResource("testfiles/tsv/ADD_PARTICIPANTS.txt")
-
-
-  "Orchestration" - {
-
-    "should import a TSV asynchronously" - {
-      "for the owner of a workspace" in {
-        implicit val token: AuthToken = ownerAuthToken
-
-        withCleanBillingProject(owner) { projectName =>
-          withWorkspace(projectName, prependUUID("owner-tsv-import")) { workspaceName =>
-
-            // call import as owner
-            val postResponse: String = Orchestration.postRequest(s"${workspaceUrl(projectName, workspaceName)}/flexibleImportEntities", testPayload)
-            // expect to get exactly one jobId back
-            val importJobIdValues: Seq[JsValue] = postResponse.parseJson.asJsObject.getFields("jobId")
-            importJobIdValues should have size 1
-
-            val importJobId: String = importJobIdValues.head match {
-              case js: JsString => js.value
-              case x => fail("got in invalid jobId: " + x.toString())
-            }
-
-            // poll for completion as owner
-            eventually {
-              val resp: HttpResponse = Orchestration.getRequest(s"${workspaceUrl(projectName, workspaceName)}/importPFB/$importJobId")
-              resp.status shouldBe StatusCodes.OK
-              blockForStringBody(resp).parseJson.asJsObject.fields.get("status").value shouldBe JsString("Done")
-            }
-
-            // inspect data entities and confirm correct import as owner
-            eventually {
-              val resp = Orchestration.getRequest(s"${ServiceTestConfig.FireCloud.orchApiUrl}api/workspaces/$projectName/$workspaceName/entities")
-              compareMetadata(blockForStringBody(resp).parseJson.convertTo[Map[String, EntityTypeMetadata]], testPayload)
-            }
-
-          }
-        }
-      }
-    }
-  }
-
-}
